+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 2.0a13 (unreleased)
 -------------------
 
+- In the event listing, use the collection's ``item_count`` attribute to limit
+  the batch size.
+  [thet]
+
 - For the event portlet, don't cache the list of events on memoize instance,
   which creates a write transaction. Remove the caching until a solid cache key
   is found, which also works for multiple portlet instances.